S03P31B306-102 [+]ADD 좋아요기능, 헤더에서 토큰 정보 받게끔 코드 전면 수정

+++ backend/src/main/java/com/daeng/nyang/dto/AnimalLike.java
@@ -20,7 +20,7 @@ public class AnimalLike {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY) // 얘가 auto_increment
-	private Long no;
+	private int no;
 
 	@Column
 	private String user_id;
